+++ CHANGELOG.md
@@ -24,6 +24,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-2.5.0...master
 
+### Fixed
+- Support `SELECT FOR NO KEY UPDATE` statement
+- Support `SELECT FOR SHARE` statement
+- Support `SELECT FOR KEY SHARE` statement
+
 ## [2.5.0] - 2021-10-25
 [2.5.0]: https://github.com/atlassian-labs/db-replica/compare/release-2.4.0...release-2.5.0